+++ Refactoring/app.js
@@ -6,6 +6,7 @@ const mongodbStore = require('connect-mongodb-session');
 const csrf = require('csurf');
 
 const db = require('./data/database');
+const authRoutes = require('./routes/auth');
 const blogRoutes = require('./routes/blog');
 
 const MongoDBStore = mongodbStore(session);
@@ -49,6 +50,7 @@ app.use(async function(req, res, next) {
 });
 
 app.use(blogRoutes);
+app.use(authRoutes);
 
 app.use(function(error, req, res, next) {
   res.render('500');
